Add `get_move_str()` function to display more info
A new function, `get_move_str()`, has been implemented to display more information for each move. Specifically, it combines information of what move number it was, whose turn it was, and what move was played in the previous position into a string. Previously, only the last of the three pieces of information would be displayed during the execution of the main program.