Improve data-collection instructions
During image-data collection, the orientation of the physical chessboard (in the perspective-transformed image) must match the orientation of the digital chessboard (in the "Current position" window). This is assumed in the subsequent data-processing steps.

Note that there are two ways to accomplish this. The obvious way is to physically rotate the chessboard and tune the slider values as usual to achieve the desired orientation. The other way is to tune the slider values unconventionally such that the `x_TL` and `y_TL` values don't necessarily correspond to the physical top-left corner, `x_TR` and `y_TR` values don't necessarily correspond to the physical top-right corner, etc.